Refactor: Split main.py into modular package architecture
Split the monolithic main.py (632 lines) into a well-organized flet_from_json package:

- flet_from_json/__init__.py: Package exports
- flet_from_json/state.py: AppState class for field values and visibility
- flet_from_json/constants.py: Color, font size, and icon mappings
- flet_from_json/elements.py: ElementBuilder class for creating UI controls
- flet_from_json/builder.py: JsonUIBuilder class for constructing the UI
- flet_from_json/config.py: Configuration loading utilities

The main.py is now a minimal 62-line entry point that imports from the package.

Updated README.md to document the new project structure.
